remove Identity and jwt token

+++ JustProject.Domain/ViewModels/AccountViewModel.cs
@@ -11,6 +11,6 @@ namespace JustProject.Domain.ViewModels
         public string FullName { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-        public int Phone { get; set; } 
+        public decimal Phone { get; set; } 
     }
 }
